Rename blockstates module
